+++ sql/sql_plugin.cc
@@ -586,8 +586,7 @@ static my_bool read_mysql_plugin_info(struct st_plugin_dl *plugin_dl,
                     MYF(MY_ZEROFILL|MY_WME));
     if (!cur)
     {
-      my_error(ER_OUTOFMEMORY, MyFlags,
-                   static_cast<int>(plugin_dl->dl.length));
+      my_error(ER_OUTOFMEMORY, MyFlags, plugin_dl->dl.length);
       DBUG_RETURN(TRUE);
     }
     /*
@@ -706,8 +705,7 @@ static my_bool read_maria_plugin_info(struct st_plugin_dl *plugin_dl,
                   MYF(MY_ZEROFILL|MY_WME));
       if (!cur)
       {
-        my_error(ER_OUTOFMEMORY, MyFlags,
-                     static_cast<int>(plugin_dl->dl.length));
+        my_error(ER_OUTOFMEMORY, MyFlags, plugin_dl->dl.length);
         DBUG_RETURN(TRUE);
       }
       /*
@@ -843,8 +841,7 @@ static st_plugin_dl *plugin_dl_add(const LEX_CSTRING *dl, myf MyFlags)
   if (! (plugin_dl.dl.str= (char*) my_malloc(key_memory_mysql_plugin_dl,
                                              plugin_dl.dl.length, MYF(0))))
   {
-    my_error(ER_OUTOFMEMORY, MyFlags,
-                 static_cast<int>(plugin_dl.dl.length));
+    my_error(ER_OUTOFMEMORY, MyFlags, plugin_dl.dl.length);
     goto ret;
   }
   plugin_dl.dl.length= copy_and_convert((char*) plugin_dl.dl.str,
@@ -856,8 +853,7 @@ static st_plugin_dl *plugin_dl_add(const LEX_CSTRING *dl, myf MyFlags)
   /* Add this dll to array */
   if (! (tmp= plugin_dl_insert_or_reuse(&plugin_dl)))
   {
-    my_error(ER_OUTOFMEMORY, MyFlags,
-                 static_cast<int>(sizeof(struct st_plugin_dl)));
+    my_error(ER_OUTOFMEMORY, MyFlags, sizeof(struct st_plugin_dl));
     goto ret;
   }
